Support for issuing client certificates
Added options for defining 3rd party clients and issue client
certificates based on pre-defined roles.

Client roles provides an option to define a set of subject elements
that will be added to the client certificate. If any of these are already
defined in a supplied CSR, the role subject elements will take precedence.

A client is associated with a specific CA, and all issued client
certificates will be signed with this CA.

Both options for issuing a certificate with a private key and to sign a
provided CSR are provided.